integer fixes, sounds and texts collected

+++ neo/idlib/containers/Sort.h
@@ -126,17 +126,17 @@ public:
 			return;
 		}
 
-		const int64_t MAX_LEVELS = 128;
-		int64_t lo[MAX_LEVELS], hi[MAX_LEVELS];
+		const int MAX_LEVELS = 128;
+		int lo[MAX_LEVELS], hi[MAX_LEVELS];
 
 		// 'lo' is the lower index, 'hi' is the upper index
 		// of the region of the array that is being sorted.
 		lo[0] = 0;
 		hi[0] = num - 1;
 
-		for ( int64_t level = 0; level >= 0; ) {
-			int64_t i = lo[level];
-			int64_t j = hi[level];
+		for ( int level = 0; level >= 0; ) {
+			int i = lo[level];
+			int j = hi[level];
 
 			// Only use quick-sort when there are 4 or more elements in this region and we are below MAX_LEVELS.
 			// Otherwise fall back to an insertion-sort.
@@ -145,7 +145,7 @@ public:
 				// Use the center element as the pivot.
 				// The median of a multi point sample could be used
 				// but simply taking the center works quite well.
-				int64_t pi = ( i + j ) / 2;
+				int pi = ( i + j ) / 2;
 
 				// Move the pivot element to the end of the region.
 				SwapValues( base[j], base[pi] );
@@ -180,8 +180,8 @@ public:
 
 				// Insertion-sort of the remaining elements.
 				for( ; i < j; j-- ) {
-					int64_t m = i;
-					for ( int64_t k = i + 1; k <= j; k++ ) {
+					int m = i;
+					for ( int k = i + 1; k <= j; k++ ) {
 						if ( static_cast< const _derived_ * >( this )->Compare( base[k], base[m] ) > 0 ) {
 							m = k;
 						}
